Changed create ride button to show near map
